fix(release): resolve RP-2 dry-run blockers (#8579)
Blocker 1 — clippy --all-targets in bench/test code:
- perl-incremental-parsing bench: add #![allow(dead_code,
  clippy::expect_used, clippy::manual_range_contains)] at file level.
  Bench scaffolding expect() is appropriate for setup failures.
- perl-module/tests/module_resolution_path_fuzz.rs: remove redundant
  `as u8` cast (byte is already u8).
- perl-module/tests/resolution_uri_comprehensive_unit_tests.rs: replace
  `&[workspace_uri.clone()]` with `std::slice::from_ref(&workspace_uri)`
  to fix cloned_ref_to_slice_refs lint.
- perl-tdd-support/tests/test_helper_coverage.rs: add
  #[allow(unused_must_use)] for unit-return must_some/must_err tests;
  use `let _ =` for non-unit must_use suppressions inside catch_unwind.

Blocker 2 — cargo-semver-checks rustdoc v57 incompatibility:
- Pin cargo-semver-checks to 0.47.0 (supports v57) in justfile
  _semver-check-install and ci-nightly.yml install step. Previous
  0.45.0 did not support the rustdoc JSON format emitted by Rust 1.95.

Third blocker discovered: perl-lsp-perltidy tests/subprocess_tests.rs
has 5 clippy::expect_used violations missed in the original receipt.
Filed as #8720 for a separate fix PR.

Updated docs/release/0.14.0/dry-run-receipt.md to reflect resolutions.

+++ crates/perl-module/tests/module_resolution_path_fuzz.rs
@@ -14,7 +14,7 @@ fn fuzz_segment(state: &mut u64, max_len: usize) -> String {
     for _ in 0..len {
         let byte = (next_u64(state) & 0x7F) as u8;
         let ch = match byte {
-            0..=31 => b'a' + (byte % 26) as u8,
+            0..=31 => b'a' + byte % 26,
             b':' | b'/' | b'\\' => b'x',
             _ => byte,
         } as char;
